Change Peer client's port parameter to 443 and project structure
While deploying on heroku, the default port for the server
is 443. So we need to make our Peer Clients connect to the Peer
server on this port only.

Also change the backend's package.json file to make it
work with heroku deployments.

I've changed the project structure to make it more suitable for
deploying to heroku. frontend directory has been included within
the backend itself so that when pushing to heroku, it builds the
node server using nodejs webpack and serves the pages in frontend's
public directory which consists of compiled/build react files.